fix(assignments): correct method renaming and prevent duplicate assignment clears

+++ app/controllers/api/v2/assignments_controller.rb
@@ -5,9 +5,9 @@ module Api
     class AssignmentsController < AnsibleDirectorApiController
       before_action :find_resource, only: %i[destroy]
       before_action :find_resources, only: %i[assign]
-      before_action :find_target, only: %i[find_assignments]
+      before_action :find_target, only: %i[get_assignments]
 
-      def find_assignments
+      def get_assignments
         @assignments = @target.resolved_ansible_content
         a = 2
       end
@@ -18,6 +18,7 @@ module Api
 
       def assign_bulk
         assignments = bulk_assignment_params
+        cleared_targets = []
         ActiveRecord::Base.transaction do
           assignments.each do |assignment|
             source_finder = finder(assignment[:source][:type])
@@ -25,7 +26,10 @@ module Api
             source = source_finder.find_by(id: assignment[:source][:id])
             target = target_finder.find_by(id: assignment[:target][:id])
 
-            AnsibleContentAssignment.where(assignable: target).destroy_all
+            unless target.id.in?(cleared_targets)
+              AnsibleContentAssignment.where(assignable: target).destroy_all
+              cleared_targets.push(target.id)
+            end
             AnsibleContentAssignment.create!(consumable: source, assignable: target)
           end
         end
@@ -64,7 +68,7 @@ module Api
         case type
 
         when 'ACR'
-          # @assignment_class = AnsibleContentAssignmentCollectionRole
+          #@assignment_class = AnsibleContentAssignmentCollectionRole
           AnsibleCollectionRole
         when 'CONTENT'
           ContentUnitVersion
@@ -85,16 +89,12 @@ module Api
         source = source_finder.find_by(id: assignment[:source][:id])
         target = target_finder.find_by(id: assignment[:target][:id])
         if source.nil?
-          message = "Source object of type #{assignment[:source][:type]} " \
-          "with id #{assignment[:source][:id]} does not exist."
           render_error('custom_error', status: :unprocessable_entity,
-                       locals: { message: message })
+                       locals: { message: "Source object of type #{assignment[:source][:type]} with id #{assignment[:source][:id]} does not exist." })
         end
         if target.nil?
-          message = "Target object of type #{assignment[:target][:type]} " \
-            "with id #{assignment[:target][:id]} does not exist."
           render_error('custom_error', status: :unprocessable_entity,
-                       locals: { message: message })
+                       locals: { message: "Target object of type #{assignment[:target][:type]} with id #{assignment[:target][:id]} does not exist." })
         end
         @source = source
         @target = target
